+++ ssb_parser/src/types/error.rs
@@ -18,7 +18,7 @@ pub struct ParseError {
 }
 impl ParseError {
     /// New error with message only.
-    pub fn new(msg: &str) -> Self {
+    pub(crate) fn new(msg: &str) -> Self {
         Self {
             msg: msg.to_owned(),
             pos: None,
@@ -26,7 +26,7 @@ impl ParseError {
         }
     }
     /// New error with message and position.
-    pub fn new_with_pos(msg: &str, pos: (usize, usize)) -> Self {
+    pub(crate) fn new_with_pos(msg: &str, pos: (usize, usize)) -> Self {
         Self {
             msg: msg.to_owned(),
             pos: Some(pos),
@@ -34,7 +34,7 @@ impl ParseError {
         }
     }
     /// New error with message and source error.
-    pub fn new_with_source<E>(msg: &str, src: E) -> Self
+    pub(crate) fn new_with_source<E>(msg: &str, src: E) -> Self
         where E: Error + 'static {
         Self {
             msg: msg.to_owned(),
@@ -43,7 +43,7 @@ impl ParseError {
         }
     }
     /// New error with message, position and source error.
-    pub fn new_with_pos_source<E>(msg: &str, pos: (usize, usize), src: E) -> Self
+    pub(crate) fn new_with_pos_source<E>(msg: &str, pos: (usize, usize), src: E) -> Self
         where E: Error + 'static {
         Self {
             msg: msg.to_owned(),